feat(examples): Add Captain-backed entity demo serving and tests
Extract the entity demo into a reusable internal package and add focused entrypoint and Captain configuration coverage.
Expose executor-backed OpenAPI, MCP, embedded UI, Vite development mode, and Captain-configured chat from one registration set.
BREAKING CHANGE: Chat no longer forces manual tool approval; execution now follows Captain chat service behavior.
