+++ examples/enitity/internal/entitydemo/entitydemo_suite_test.go
@@ -1,0 +1,13 @@
+package entitydemo
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestEntityDemo(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Entity Demo Suite")
+}
