+++ examples/enitity/internal/entitydemo/entitydemo_test.go
@@ -1,4 +1,4 @@
-package main
+package entitydemo
 
 import (
 	"encoding/json"
